gitvibe-backend/pyproject.toml
[tool.poetry]
name = "gitvibe-backend"
version = "0.1.0"
description = "GitVibe Backend - Analyze GitHub repositories with style and humor"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.24.0" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.6"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
redis = "^5.0.1"
aioredis = "^2.0.1"
openai = "^1.82.0"
python-gitlab = "^3.15.0"
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
